setup react component with redux
