+++ Procfile.dev
@@ -1,2 +1,2 @@
-web: bin/rails server -p 3000
+web: bundle exec rails s
 js: yarn build --watch
