+++ app/javascript/app/app.js
@@ -1,0 +1,23 @@
+import React from 'react';
+import { Routes, Route } from 'react-router-dom';
+import { BrowserRouter } from "react-router-dom";
+import { Provider } from "react-redux";
+import store from "./store";
+import Message from '../feature/message/message';
+
+function App() {
+
+  return (
+    <React.StrictMode>
+      <Provider store={store}>
+        <BrowserRouter>
+          <Routes>
+            <Route path="/" element={<Message />} />
+          </Routes>
+        </BrowserRouter>
+      </Provider>
+    </React.StrictMode>
+  );
+}
+
+export default App;
