--- app/javascript/app/app.jsx
@@ -1,20 +1,0 @@
-import React, { useEffect } from 'react';
-import { useDispatch } from 'react-redux';
-import { Routes, Route } from 'react-router-dom';
-import { fetchMessage } from '../feature/message/messageSlice';
-
-function App() {
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    dispatch(fetchMessage());
-  }, [dispatch]);
-
-  return (
-    <Routes>
-      <Route path="/" element={<Greeting />} />
-    </Routes>
-  );
-}
-
-export default App;
