+++ app/javascript/application.js
@@ -1,3 +1,7 @@
 // Entry point for the build script in your package.json
-import "@hotwired/turbo-rails"
-import "./controllers"
+import React from "react";
+import { createRoot } from 'react-dom/client'
+import App from "./app/app";
+
+const root = createRoot(document.getElementById('root'));
+root.render(<App />);
