+++ app/javascript/feature/message/messageSlice.js
@@ -4,9 +4,9 @@ import { createSlice } from "@reduxjs/toolkit";
 export const fetchMessage = createAsyncThunk(
   "messages/fetchMessages",
   async () => {
-    const response = await fetch("/api/messages");
-    const { message } = await response.json();
-    return message;
+    const response = await fetch("/api/greetings");
+    const message = await response.json();
+    return message.message;
   }
 );
 
@@ -18,7 +18,8 @@ export const messageSlice = createSlice({
   reducers: {},
   extraReducers: {
     [fetchMessage.fulfilled]: (state, action) => {
-      state.message = action.payload;
+      const newState = { message: action.payload };
+      return newState;
     }
   }
 });
